Packaging: first try with poetry
TODO
- Fix path deps
- Make application factories pattern usable:
https://flask.palletsprojects.com/en/1.1.x/extensiondev/
