+++ .gitignore
@@ -1,2 +1,3 @@
 docs/build/
 docs/source/_build/
+dist/
