+++ README.md
@@ -38,3 +38,15 @@ This code represent the idea behind the module, it's not real code, look at app.
 ## Contribute : Pull requests are welcome !
 
 [![Edit with Gitpod](https://gitpod.io/button/open-in-gitpod.svg)](https://gitpod.io/#https://github.com/playerla/flask-socketio-lit-html/tree/Dev)
+
+### Updating autodoc
+
+```sh
+cd docs && sphinx-apidoc -o source/ ../flask_socketio_lit_html
+```
+
+### Build package
+
+```sh
+poetry build
+```
