--- docs/source/app.rst
@@ -1,7 +1,0 @@
-app module
-==========
-
-.. automodule:: app
-   :members:
-   :undoc-members:
-   :show-inheritance:
