+++ docs/source/modules.rst
@@ -1,0 +1,7 @@
+flask_socketio_lit_html
+=======================
+
+.. toctree::
+   :maxdepth: 4
+
+   webcomponent_base
