+++ docs/source/webcomponent_base.rst
@@ -1,0 +1,7 @@
+webcomponent\_base module
+=========================
+
+.. automodule:: webcomponent_base
+   :members:
+   :undoc-members:
+   :show-inheritance:
