+++ flask_socketio_lit_html/templates/main.html
