+++ flask_socketio_lit_html/webcomponent_templates/webcomponent_base.js
