+++ pyproject.toml
@@ -1,0 +1,15 @@
+[tool.poetry]
+name = "flask-socketio-lit-html"
+version = "0.1.0"
+description = "Simple Webcomponents with flask"
+authors = ["playerla <playerla.94@gmail.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.7"
+
+[tool.poetry.dev-dependencies]
+pytest = "^3.0"
+
+[build-system]
+requires = ["poetry>=0.12"]
+build-backend = "poetry.masonry.api"
